requirements.txt
requests
beautifulsoup4
google-generativeai
dotenv